Split main.py into logical parts
Added new encode.py and decode.py files, changed .gitignore accordingly
