+++ .gitignore
@@ -1,5 +1,6 @@
 *
 
 !main.py
+!*code.py
 !.gitignore
 !README.md
